Flytter all kode i MittNavVarselService.kt til KjørelistevarselService.kt da all kode der omhandler logikk rundt kjøreliste-varsling
